% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel_all_orders.R
\name{cancel_all_orders}
\alias{cancel_all_orders}
\title{Cancel all order(s) for a symbol}
\usage{
cancel_all_orders(symbols = NULL, tradeType = NULL, delay = 0, retries = 3)
}
\arguments{
\item{symbols}{A `character` vector of one or more symbols; format "BTC/USDT" (optional - default `NULL`).}

\item{tradeType}{A `character` vector of one either "TRADE" or "MARGIN_ISOLATED_TRADE" (optional - default `NULL`).}

\item{delay}{A `numeric` value to delay data request in milliseconds (optional - default `0`).}

\item{retries}{A `numeric` value to specify the number of retries in case of failure (optional - default `3`).}
}
\value{
Success returns `character` vector of n order id(s) (designated by KuCoin).
}
\description{
TODO: in development.
}
\details{
This endpoint requires the Trade permission.

This API is restricted for each account, the request rate limit is 3 times/3s.

# ---------------
For more information see documentation: [KuCoin - cancel-all-orders](https://docs.kucoin.com/#cancel-all-orders)

TODO: if no symbols provided cancel all
}
\examples{

\dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

symbol <- "ETH/USDT"

# check balances
balances <- kucoin::get_kucoin_balances(); balances

order_id <- kucoin::post_kucoin_limit_order(
    symbol = symbol,
    side = "sell",
    base_size = 1,
    price = 1000000
); order_id

order_id2 <- kucoin::post_kucoin_limit_order(
    symbol = symbol,
    side = "sell",
    base_size = 1,
    price = 1000000
); order_id

# see funds on hold
balances2 <- kucoin::get_kucoin_balances(); balances2

# cancel all orders for symbol
cancelled <- kucoin::cancel_all_orders(symbol); cancelled

# cancel all orders for all symbols and trade types
cancelled <- kucoin::cancel_all_orders(); cancelled

# see funds released
balances3 <- kucoin::get_kucoin_balances(); balances3

}

}
