% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel-all-orders.R
\name{cancel_kucoin_all_pairs_order}
\alias{cancel_kucoin_all_pairs_order}
\title{Cancel all order(s) for a symbol}
\usage{
cancel_kucoin_all_pairs_order(
  symbols,
  tradeType = "TRADE",
  delay = 0,
  retries = 3
)
}
\arguments{
\item{symbols}{A \code{character} vector of one or more symbols; format "BTC/USDT".}

\item{tradeType}{A \code{character} vector of one either "TRADE" or "MARGIN_ISOLATED_TRADE".}

\item{delay}{A \code{numeric} value to delay data request in milliseconds.}

\item{retries}{A \code{numeric} value to specify the number of retries in case of failure.}
}
\value{
If success returns \code{character} vector of n order id(s); designated by KuCoin.
}
\description{
Cancel all order(s) for a symbol
}
\details{
This API is restricted for each account, the request rate limit is 3 times/3s.

For more information see documentation: \href{https://docs.kucoin.com/#cancel-all-orders}{KuCoin - cancel-all-orders}
}
\examples{

\dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

# import library
library("kucoin")

symbol <- "ETH/USDT"

# check balances
balances <- kucoin::get_kucoin_balances(); balances

order_id <- kucoin::post_kucoin_limit_order(
    symbol = symbol,
    side = "sell",
    base_size = 1,
    price = 1000000
); order_id

order_id2 <- kucoin::post_kucoin_limit_order(
    symbol = symbol,
    side = "sell",
    base_size = 1,
    price = 1000000
); order_id

# see funds on hold
balances2 <- kucoin::get_kucoin_balances(); balances2

# cancel all orders for symbol
cancelled <- kucoin::cancel_kucoin_all_pairs_order(symbol); cancelled

# see funds released
balances3 <- kucoin::get_kucoin_balances(); balances3

}

}
