% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel-order.R
\name{cancel_kucoin_order}
\alias{cancel_kucoin_order}
\title{Cancel order(s)}
\usage{
cancel_kucoin_order(orderId, delay = 0, retries = 3)
}
\arguments{
\item{delay}{A \code{numeric} value to delay data request in milliseconds.}

\item{retries}{A \code{numeric} value to specify the number of retries in case of failure.}

\item{order_ids}{A \code{character} vector of one or more which contain the order id(s) designated by KuCoin.}
}
\value{
If success returns \code{character} vector of n order id(s); designated by KuCoin.
}
\description{
Cancel order(s)
}
\details{
This API is restricted for each account, the request rate limit is 60 times/3s.

For more information see documentation: \href{https://docs.kucoin.com/#cancel-an-order}{KuCoin - cancel-order}
}
\examples{

\dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

# import library
library("kucoin")

# set a limit order
order_id <- kucoin::post_kucoin_limit_order(
   symbol = "ETH/USDT",
   side = "sell",
   base_size = 1,
   price = 100000
)

# get order details
cancelled <- kucoin::cancel_kucoin_order(order_id)

}

}
