% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl_market_data_get_klines.R
\docType{data}
\name{freq_to_second_map}
\alias{freq_to_second_map}
\title{Get Klines market data allowed frequencies}
\format{
An object of class \code{list} of length 14.
}
\usage{
freq_to_second_map
}
\description{
Get Klines market data allowed frequencies
}
\keyword{datasets}
