% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl_market_data_get_klines.R
\name{frequency_to_seconds}
\alias{frequency_to_seconds}
\title{Convert Frequency String to Seconds}
\usage{
frequency_to_seconds(freq)
}
\arguments{
\item{freq}{A character string representing the frequency. Allowed values are:
"1min", "3min", "5min", "15min", "30min", "1hour", "2hour", "4hour", "6hour", "8hour",
"12hour", "1day", "1week", "1month".}
}
\value{
A numeric value representing the duration in seconds.
}
\description{
Given a frequency string (e.g. "1min", "3min", "1hour", etc.), this function returns
the corresponding duration in seconds.
}
\examples{
frequency_to_seconds("1min")  # returns 60

}
