% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_balances.R
\name{get_account_balances}
\alias{get_account_balances}
\title{Get user's balance(s) list}
\usage{
get_account_balances(currency = NULL, type = NULL)
}
\arguments{
\item{currency}{A `character` vector of one currency symbol (optional).}

\item{type}{A `character` vector of one indicating the `"main"` or `"trade"` account type (optional).}
}
\value{
A `data.table` containing balance details
}
\description{
Get user's balance(s) list
}
\examples{

\dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

# get user's balance details
kucoin::get_account_balances()

# get user's balance details for BTC only
kucoin::get_account_balances(currency = "BTC")

# get user's balance details for trade account only
kucoin::get_account_balances(type = "trade")

}

}
