% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deposit_address.R
\name{get_deposit_address}
\alias{get_deposit_address}
\title{Get a deposit address for a currency}
\usage{
get_deposit_address(currency = NULL, chain = NULL)
}
\arguments{
\item{currency}{A \code{character} vector length 1 to specify the currencies to get deposit addresses for (required - default \code{NULL})}

\item{chain}{A \code{character} vector length 1 to specify the chain (network) to use for the deposit address (required - default \code{NULL}).}
}
\value{
A \code{data.table} containing deposit address information
}
\description{
Get a deposit address for a currency. This function is useful for generating a deposit address for a currency. Note you must provide the chain (network) to use for the deposit address. This can be obtained by use of \code{\link[=get_currency_details]{get_currency_details()}}.
}
\details{
For more information see documentation: \href{https://docs.kucoin.com/#get-deposit-addresses-v2}{KucCoin - get-deposit-addresses-v2}.
}
\examples{

\dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

# import library
library("kucoin")

# get a deposit address for a currency
deposit_address <- kucoin::get_deposit_address("BTC", "btc")

# quick check
deposit_address

}

}
\seealso{
\code{kucoin::get_currency_details()()}
}
