% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_futures_base_url}
\alias{get_futures_base_url}
\title{Retrieve KuCoin Futures API Base URL}
\usage{
get_futures_base_url(url = Sys.getenv("KUCOIN_FUTURES_API_ENDPOINT"))
}
\arguments{
\item{url}{(scalar<character> | NULL) explicit base URL. Defaults to
\code{Sys.getenv("KUCOIN_FUTURES_API_ENDPOINT")}.}
}
\value{
(scalar<character>) the Futures API base URL.
}
\description{
Returns the base URL for the KuCoin Futures API in the following priority:
\enumerate{
\item The explicitly provided \code{url} parameter.
\item The \code{KUCOIN_FUTURES_API_ENDPOINT} environment variable.
\item The default \code{"https://api-futures.kucoin.com"}.
}
}
\examples{
\dontrun{
get_futures_base_url()
}
}
