% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-info.R
\name{get_kucoin_balances}
\alias{get_kucoin_balances}
\title{Get user's balance(s) list}
\usage{
get_kucoin_balances(currency = NULL, type = NULL)
}
\arguments{
\item{currency}{A \code{character} vector of one currency symbol (optional).}

\item{type}{A \code{character} vector of one indicating the \code{"main"} or \code{"trade"} account type (optional).}
}
\value{
A \code{data.table} containing balance details
}
\description{
Get user's balance(s) list
}
\examples{

\dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

# import library
library("kucoin")

# get user's balance details
balances <- get_kucoin_balances()

# quick check
balances

# get user's balance details for BTC only
balances <- get_kucoin_balances(
  currency = "BTC"
)

# quick check
balances

# get user's balance details for trade account only
balances <- get_kucoin_balances(
  type = "trade"
)

# quick check
balances

}

}
