% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency-data.R
\name{get_kucoin_currencies_details}
\alias{get_kucoin_currencies_details}
\title{Get a currencies' details}
\usage{
get_kucoin_currencies_details(currencies)
}
\arguments{
\item{currencies}{A \code{character} vector to specify the currencies to get details for.}
}
\value{
A \code{data.table} containing currency information
}
\description{
Get a currencies' details. This includes what chains are available for depositing; this function is useful for then generating a deposit address by use of \code{kucoin::get_kucoin_deposit_address()}.
}
\examples{
# import library
library("kucoin")

# get a currencies' details
currencies_details <- get_kucoin_currencies_details(c("BTC", "XMR"))

# quick check
currencies_details

}
\seealso{
\code{kucoin::get_kucoin_deposit_address()}
}
