% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency-deposit.R
\name{get_kucoin_deposit_address}
\alias{get_kucoin_deposit_address}
\title{Get a deposit address for a currency}
\usage{
get_kucoin_deposit_address(currency, chain = NULL)
}
\arguments{
\item{chain}{A \code{character} vector length 1 to specify the chain (network) to use for the deposit address.}

\item{currencies}{A \code{character} vector length 1 to specify the currencies to get deposit addresses for.}
}
\value{
A \code{data.table} containing deposit address information
}
\description{
Get a deposit address for a currency. This function is useful for generating a deposit address for a currency. Note you must provide the chain (network) to use for the deposit address. This can be obtained by use of \code{kucoin::get_kucoin_currencies_details()}.
}
\examples{

\dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

# import library
library("kucoin")

# get a deposit address for a currency
deposit_address <- get_kucoin_deposit_address("BTC", "btc")

# quick check
deposit_address

}

}
\seealso{
\code{kucoin::get_kucoin_currencies_details()}
}
