% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-order.R
\name{get_kucoin_order}
\alias{get_kucoin_order}
\title{Get an order details}
\usage{
get_kucoin_order(order_ids, delay = 0, retries = 3)
}
\arguments{
\item{order_ids}{A \code{character} vector of one or more which contain the order id(s) designated by KuCoin.}

\item{delay}{A \code{numeric} value to delay data request in milliseconds.}

\item{retries}{A \code{numeric} value to specify the number of retries in case of failure.}
}
\value{
A \code{data.table} containing order details
}
\description{
Get an order details
}
\examples{

\dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

# import library
library("kucoin")

# get order details
order_details <- get_kucoin_order(
  order_ids = "insertorderid"
)

# quick check
order_details

}

}
