% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market-data.R
\name{get_kucoin_prices}
\alias{get_kucoin_prices}
\title{Get historical data from specified symbols}
\usage{
get_kucoin_prices(symbols, from, to, frequency, delay = 0, retries = 3)
}
\arguments{
\item{symbols}{A \code{character} vector of one or more pair symbol.}

\item{from}{A \code{character} with valid \code{date}/\code{datetime} format, or \code{date}/\code{datetime} object as a start of datetime range.}

\item{to}{A \code{character} with valid \code{date}/\code{datetime} format, or \code{date}/\code{datetime} object as an end of datetime range.}

\item{frequency}{A \code{character} vector of one which specify the frequency option, see details for further information.}

\item{delay}{A \code{numeric} value to delay data request in milliseconds.}

\item{retries}{A \code{numeric} value to specify the number of retries in case of failure.}
}
\value{
A \code{data.table} containing prices data
}
\description{
Get historical data from specified symbols
}
\details{
There are several supported frequencies:
\itemize{
\item \code{"1 minute"}
\item \code{"3 minutes"}
\item \code{"5 minutes"}
\item \code{"15 minutes"}
\item \code{"30 minutes"}
\item \code{"1 hour"}
\item \code{"2 hours"}
\item \code{"4 hours"}
\item \code{"6 hours"}
\item \code{"8 hours"}
\item \code{"12 hours"}
\item \code{"1 day"}
\item \code{"1 week"}
}
}
\examples{

# import library
library("kucoin")

# get one pair of symbol prices
prices <- get_kucoin_prices(
  symbols = "KCS/USDT",
  from = "2022-11-05 00:00:00",
  to = "2022-11-06 00:00:00",
  frequency = "1 hour"
)

# quick check
prices

# get multiple pair of symbols prices
prices <- get_kucoin_prices(
  symbols = c("KCS/USDT", "BTC/USDT", "KCS/BTC"),
  from = "2022-11-05 00:00:00",
  to = "2022-11-06 00:00:00",
  frequency = "1 hour"
)

# quick check
prices

}
