% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_kucoin_time}
\alias{get_kucoin_time}
\title{Get current KuCoin API server time}
\usage{
get_kucoin_time(raw = FALSE)
}
\arguments{
\item{raw}{A \code{logical} vector to specify whether
to return a raw results or not. The default is \code{FALSE}.}
}
\value{
A \code{datetime} object
}
\description{
Get current KuCoin API server time
}
\examples{
# import library
library(kucoin)

# get current server time
get_kucoin_time()

}
