% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_data.R
\name{get_market_data}
\alias{get_market_data}
\title{Get historical market data for symbols}
\usage{
get_market_data(
  symbols = NULL,
  from = NULL,
  to = NULL,
  frequency = NULL,
  delay = 0,
  retries = 3
)
}
\arguments{
\item{symbols}{A \code{character} vector of one or more pair symbol (required - default \code{NULL}).}

\item{from}{A \code{character} with valid \code{date}/\code{datetime} format, or \code{date}/\code{datetime} object as a start of datetime range (required - default \code{NULL}).}

\item{to}{A \code{character} with valid \code{date}/\code{datetime} format, or \code{date}/\code{datetime} object as an end of datetime range (required - default \code{NULL}).}

\item{frequency}{A \code{character} vector of one which specify the frequency option, see details for further information (required - default \code{NULL}).}

\item{delay}{A \code{numeric} value to delay data request in milliseconds (optional - default \code{0}).}

\item{retries}{A \code{numeric} value to specify the number of retries in case of failure (optional - default \code{3}).}
}
\value{
A \code{data.table} with price data
}
\description{
Get historical market data for symbols
}
\details{
There are several supported frequencies:
\itemize{
\item \code{"1 minute"}
\item \code{"3 minutes"}
\item \code{"5 minutes"}
\item \code{"15 minutes"}
\item \code{"30 minutes"}
\item \code{"1 hour"}
\item \code{"2 hours"}
\item \code{"4 hours"}
\item \code{"6 hours"}
\item \code{"8 hours"}
\item \code{"12 hours"}
\item \code{"1 day"}
\item \code{"1 week"}
}
}
\section{---------------}{
For more information see documentation: \href{https://docs.kucoin.com/#get-klines}{KuCoin - get-klines}
}

\examples{

# get one symbol
kucoin::get_market_data(
    symbols = "BTC/USDT",
    from = "2022-11-05 00:00:00",
    to = "2022-11-06 00:00:00",
    frequency = "1 hour"
)

# get multiple symbols
kucoin::get_market_data(
    symbols = c("BTC/USDT", "XMR/BTC", "KCS/USDT"),
    from = "2022-11-05 00:00:00",
    to = "2022-11-06 00:00:00",
    frequency = "1 hour"
)

}
