% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_metadata.R
\name{get_market_metadata}
\alias{get_market_metadata}
\title{Get all market symbols' metadata}
\usage{
get_market_metadata(retries = 3)
}
\arguments{
\item{retries}{A `numeric` value to specify the number of retries in case of failure (optional - default `3`).}
}
\value{
A `data.table` with metadata
}
\description{
Get all market symbols' metadata
}
\details{
# ---------------
For more information see documentation: [KuCoin - get-symbols-list-deprecated](https://docs.kucoin.com/#get-symbols-list-deprecated)
}
\examples{

# get all symbols' most recent metadata
kucoin::get_market_metadata()

}
