% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orders_all.R
\name{get_orders_all}
\alias{get_orders_all}
\title{Get all order(s)}
\usage{
get_orders_all(
  symbol = NULL,
  status = NULL,
  side = NULL,
  type = NULL,
  tradeType = "TRADE",
  from = NULL,
  to = NULL,
  retries = 3,
  delay = 0.1
)
}
\arguments{
\item{symbol}{A `character` vector of one or more which contain symbol(s) of format "BTC/USDT" (optional - default `NULL`).}

\item{status}{A `character` vector of one either "done" or "active" (optional - default `NULL`).}

\item{side}{A `character` vector of one either "buy" or "sell" (optional - default `NULL`).}

\item{type}{A `character` vector of one either "limit", "market", "limit_stop", "market_stop" (optional - default `NULL`).}

\item{tradeType}{A `character` vector of one either "TRADE" (spot trading), "MARGIN_TRADE" (cross margin trading), "MARGIN_ISOLATED_TRADE" (isolated margin trading) (required - default `TRADE`).}

\item{from}{A `character` with valid `date`/`datetime` format, or `date`/`datetime` object as a start of datetime range (optional - default `NULL`).}

\item{to}{A `character` with valid `date`/`datetime` format, or `date`/`datetime` object as an end of datetime range (optional - default `NULL`).}

\item{retries}{A `numeric` value to specify the number of retries in case of failure (optional - default `3`).}

\item{delay}{A `numeric` value to delay data request in milliseconds (optional - default `0.1`).}
}
\value{
A `data.table` containing order details
}
\description{
TODO: experimental.
}
\details{
For more information see documentation: [KuCoin - list-orders](https://docs.kucoin.com/#list-orders)
}
\examples{

\dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

kucoin::get_orders_all()

}

}
