% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orders_by_id.R
\name{get_orders_by_id}
\alias{get_orders_by_id}
\title{Get order(s) details by order id}
\usage{
get_orders_by_id(orderIds = NULL, delay = 0, retries = 3)
}
\arguments{
\item{orderIds}{A `character` vector of one or more which contain the order id(s) designated by KuCoin (required - default `NULL`).}

\item{delay}{A `numeric` value to delay data request in milliseconds (optional - default `0`).}

\item{retries}{A `numeric` value to specify the number of retries in case of failure (optional - default `3`).}
}
\value{
A `data.table` containing order details
}
\description{
Get order(s) details by order id
}
\details{
For more information see documentation: [KuCoin - get-order](https://docs.kucoin.com/#get-an-order)
}
\examples{

\dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

# get order details
order_details <- kucoin::get_orders_by_id(order_ids = "insertorderid")

# quick check
order_details

}

}
