% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kucoin.R
\docType{package}
\name{kucoin-package}
\alias{kucoin}
\alias{kucoin-package}
\title{`kucoin`: An R API to KuCoin Crytocurrency Exchange Market}
\description{
An R API to KuCoin crytocurrency exchange market.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dereckmezquita/kucoin}
}

}
\author{
\strong{Maintainer}: Dereck Mezquita \email{dereck@mezquita.io} (\href{https://orcid.org/0000-0002-9307-6762}{ORCID})

}
