% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market-order.R
\name{post_kucoin_market_order}
\alias{post_kucoin_market_order}
\title{Post a market order}
\usage{
post_kucoin_market_order(symbol, side, base_size = NULL, quote_size = NULL)
}
\arguments{
\item{symbol}{A \code{character} vector of one or more pair symbol.}

\item{side}{A \code{character} vector of one which specify the order side: \code{"buy"} or \code{"sell"}.}

\item{base_size}{A \code{numeric} vector of one determining the base size of the order.}

\item{quote_size}{A \code{numeric} vector which specify the base or quote currency size.}
}
\value{
If the transaction success, it will return a \code{character} vector which showing the order id
}
\description{
Post a market order
}
\details{
This API is restricted for each account, the request rate limit is 45 times/3s.
}
\examples{

\dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

# import library
library("kucoin")

# post a market order: buy 1 KCS
order_id <- post_kucoin_market_order(
  symbol = "KCS/BTC",
  side = "buy",
  base_size = 1
)

# quick check
order_id

# post a market order: sell 1 KCS
order_id <- post_kucoin_market_order(
  symbol = "KCS/BTC",
  side = "sell",
  base_size = 1
)

# quick check
order_id

# post a market order: buy KCS worth 0.0001 BTC
order_id <- post_kucoin_market_order(
  symbol = "KCS/BTC",
  side = "buy",
  quote_size = 0.0001
)

# quick check
order_id

# post a market order: sell KCS worth 0.0001 BTC
order_id <- post_kucoin_market_order(
  symbol = "KCS/BTC",
  side = "sell",
  base_size = 0.0001
)

# quick check
order_id

}

}
