% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_market_order.R
\name{submit_market_order}
\alias{submit_market_order}
\title{Post a market order}
\usage{
submit_market_order(
  symbol = NULL,
  side = NULL,
  base_size = NULL,
  quote_size = NULL
)
}
\arguments{
\item{symbol}{A `character` vector of one or more pair symbol (required - default `NULL`).}

\item{side}{A `character` vector of one which specify the order side: `"buy"` or `"sell"` (required - default `NULL`).}

\item{base_size}{A `numeric` vector of one determining the base size of the order; n units of the first currency in the pair (required or `quote_size` - default `NULL`).}

\item{quote_size}{A `numeric` vector which specify the base or quote currency size; n units of the second currency in the pair (required or `base_size` - default `NULL`).}
}
\value{
If success returns `character` vector of one; order id designated by KuCoin.
}
\description{
Post a market order
}
\details{
For more information see the [KuCoin API documentation - new order](https://docs.kucoin.com/#place-a-new-order).

This API is restricted for each account, the request rate limit is 45 times/3s.

Currencies are traded in pairs. The first currency is called the base currency and the second currency is called the quote currency. So for example, BTC/USDT, means that the base currency is the BTC and the quote currency is the USDT.

This function returns the order ID set by KuCoin typically looks like this: "63810a330091a60001ceeb04". This can be used to get the status of the order. See the function [kucoin::get_orders_by_id()].
}
\examples{

\dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

# check our balances
kucoin::get_account_balances()

## -------
# to avoid errors we will calculate our price and order size
# kucoin accepts these values in specific increments per ticker symbol

# get the market metadata
ticker <- "BTC/USDT"
base_to_buy <- 0.0001 # amount of btc
quote_to_spend <- 1 # 1 usdt

metadata <- kucoin::get_market_metadata.deprecated()[symbol == ticker, ]
base_size_increment <- metadata[symbol == ticker, ]$base_increment
quote_size_increment <- metadata[symbol == ticker, ]$quote_increment

# calculate size
base_size <- floor(amount_to_buy / base_size_increment) * base_size_increment
quote_size <- floor(amount_to_spend / quote_size_increment) * quote_size_increment

# post a market order: buy 1 ETH
order_id1 <- kucoin::submit_market_order(
    symbol = ticker,
    side = "buy",
    base_size = base_size
); order_id1

# post a market order: sell 1 ETH
order_id2 <- kucoin::submit_market_order(
    symbol = ticker,
    side = "sell",
    base_size = base_size
); order_id2

# post a market order: buy ETH worth 0.0001 BTC
order_id3 <- kucoin::submit_market_order(
    symbol = ticker,
    side = "buy",
    quote_size = quote_size
); order_id3

# post a market order: sell ETH worth 0.0001 BTC
order_id4 <- kucoin::submit_market_order(
    symbol = ticker,
    side = "sell",
    quote_size = quote_size
); order_id4

kucoin::get_orders_by_id(c(order_id1, order_id2, order_id3, order_id4))

}

}
