% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time.R
\name{time_convert_from_kucoin}
\alias{time_convert_from_kucoin}
\title{Convert KuCoin Timestamp to POSIXct}
\usage{
time_convert_from_kucoin(time_value, unit = c("ms", "ns", "s"))
}
\arguments{
\item{time_value}{(vector<numeric, 1..>) the UNIX timestamp(s).}

\item{unit}{(scalar<character in c("ms", "ns", "s")>) input unit: \code{"ms"}
(milliseconds, default), \code{"ns"} (nanoseconds), or \code{"s"} (seconds).}
}
\value{
(class<POSIXct>) POSIXct in UTC.
}
\description{
Converts a UNIX timestamp from KuCoin's API into a POSIXct object in UTC.
}
\examples{
\dontrun{
time_convert_from_kucoin(1698777600000, unit = "ms")
time_convert_from_kucoin(1698777600000000000, unit = "ns")
time_convert_from_kucoin(1698777600, unit = "s")
}

}
