% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel_all_orders.R
\name{cancel_all_orders}
\alias{cancel_all_orders}
\title{Cancel all order(s) for a symbol}
\usage{
cancel_all_orders(symbols = NULL, tradeType = NULL, delay = 0, retries = 3)
}
\arguments{
\item{symbols}{A \code{character} vector of one or more symbols; format "BTC/USDT" (optional - default \code{NULL}).}

\item{tradeType}{A \code{character} vector of one either "TRADE" or "MARGIN_ISOLATED_TRADE" (optional - default \code{NULL}).}

\item{delay}{A \code{numeric} value to delay data request in milliseconds (optional - default \code{0}).}

\item{retries}{A \code{numeric} value to specify the number of retries in case of failure (optional - default \code{3}).}
}
\value{
Success returns \code{character} vector of n order id(s) (designated by KuCoin).
}
\description{
Cancel all order(s) for a symbol
}
\details{
This endpoint requires the Trade permission.

This API is restricted for each account, the request rate limit is 3 times/3s.
}
\section{---------------}{
For more information see documentation: \href{https://docs.kucoin.com/#cancel-all-orders}{KuCoin - cancel-all-orders}
}

\examples{

\dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

# import library
library("kucoin")

symbol <- "ETH/USDT"

# check balances
balances <- kucoin::get_kucoin_balances(); balances

order_id <- kucoin::post_kucoin_limit_order(
    symbol = symbol,
    side = "sell",
    base_size = 1,
    price = 1000000
); order_id

order_id2 <- kucoin::post_kucoin_limit_order(
    symbol = symbol,
    side = "sell",
    base_size = 1,
    price = 1000000
); order_id

# see funds on hold
balances2 <- kucoin::get_kucoin_balances(); balances2

# cancel all orders for symbol
cancelled <- kucoin::cancel_all_orders(symbol); cancelled

# cancel all orders for all symbols and trade types
cancelled <- kucoin::cancel_all_orders(); cancelled

# see funds released
balances3 <- kucoin::get_kucoin_balances(); balances3

}

}
