% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel_order.R
\name{cancel_order}
\alias{cancel_order}
\title{Cancel order(s)}
\usage{
cancel_order(orderIds = NULL, delay = 0, retries = 3)
}
\arguments{
\item{orderIds}{A \code{character} vector of one or more which contains the order id(s) designated by KuCoin (required - default \code{NULL}).}

\item{delay}{A \code{numeric} value to delay data request in milliseconds (optional - default \code{0}).}

\item{retries}{A \code{numeric} value to specify the number of retries in case of failure (optional - default \code{3}).}
}
\value{
If success returns \code{character} vector of n order id(s); designated by KuCoin.
}
\description{
Cancel order(s)
}
\details{
This endpoint requires the Trade permission.

This API is restricted for each account, the request rate limit is 60 times/3s.

This interface is only for cancellation requests. The cancellation result needs to be obtained by querying the order status, you can use \code{\link[=get_orders_by_id]{get_orders_by_id()}}.

It is recommended that you DO NOT cancel the order until receiving the Open message, otherwise the order cannot be cancelled successfully.
}
\section{---------------}{
For more information see documentation: \href{https://docs.kucoin.com/#cancel-an-order}{KuCoin - cancel-order}
}

\examples{

\dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

# import library
library("kucoin")

# set a limit order
order_id1 <- kucoin::post_kucoin_limit_order(
   symbol = "ETH/USDT",
   side = "sell",
   base_size = 1,
   price = 100000
)

order_id2 <- kucoin::post_kucoin_limit_order(
   symbol = "ETH/USDT",
   side = "sell",
   base_size = 1,
   price = 100000
)

# get order details
cancelled <- kucoin::cancel_order(c(order_id1, order_id2))

}

}
