% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl_market_data_get_klines.R
\name{check_allowed_frequency_s}
\alias{check_allowed_frequency_s}
\title{Check Allowed Frequency in Seconds}
\usage{
check_allowed_frequency_s(freq_s)
}
\arguments{
\item{freq_s}{A numeric value representing the frequency in seconds.}
}
\description{
Check Allowed Frequency in Seconds
}
