% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl_market_data_get_klines.R
\name{check_allowed_frequency_str}
\alias{check_allowed_frequency_str}
\title{Check Allowed Frequency String}
\usage{
check_allowed_frequency_str(freq_str)
}
\arguments{
\item{freq_str}{A character string representing the frequency (e.g. "1min", "3min", etc.).}
}
\description{
Check Allowed Frequency String
}
