% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_base_url}
\alias{get_base_url}
\title{Get Base URL for KuCoin API}
\usage{
get_base_url(url = Sys.getenv("KC-API-ENDPOINT"))
}
\arguments{
\item{url}{A character string representing the base URL. Default is NULL, which triggers
the fallback behavior.}
}
\value{
A character string containing the determined base URL.
}
\description{
Get Base URL for KuCoin API
}
\details{
Returns the base URL for the KuCoin API. The URL is determined using the following priority:
1. If a URL is explicitly provided via the `url` parameter, it will be used.
2. If no URL is provided (NULL), it checks for the "KC-API-ENDPOINT" environment variable.
3. If neither source provides a URL, it falls back to the default "https://api.kucoin.com".
}
