% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_currency_details.R
\name{get_currency_details}
\alias{get_currency_details}
\title{Get currencies' details}
\usage{
get_currency_details(currencies = NULL)
}
\arguments{
\item{currencies}{A \code{character} vector to specify the currencies to get details for (required - default \code{NULL}).}
}
\value{
A \code{data.table} with currency information
}
\description{
Get a currencies' details. This includes what chains are available for depositing; this function is useful for then generating a deposit address by use of \code{\link[=get_deposit_address]{get_deposit_address()}}.\tabular{lllllllllllll}{
   currency \tab name \tab full_name \tab precision \tab is_margin_enabled \tab is_debit_enabled \tab chain_name \tab withdrawal_min_size \tab withdrawal_min_fee \tab is_withdraw_enabled \tab is_deposit_enabled \tab confirms \tab contract_address \cr
   BTC \tab BTC \tab BTC \tab 8 \tab TRUE \tab TRUE \tab BTC \tab 0.001 \tab 0.001 \tab TRUE \tab TRUE \tab 3 \tab   \cr
   BTC \tab BTC \tab BTC \tab 8 \tab TRUE \tab TRUE \tab TRC20 \tab 0.001 \tab 0.003 \tab TRUE \tab TRUE \tab 19 \tab   \cr
   BTC \tab BTC \tab BTC \tab 8 \tab TRUE \tab TRUE \tab BTC-Segwit \tab 0.001 \tab 0.001 \tab TRUE \tab TRUE \tab 3 \tab   \cr
}
}
\details{
For more information see documentation: \href{https://docs.kucoin.com/#get-currency-detail-recommend}{KuCoin - get-currency-detail}
}
\examples{
# import library
library("kucoin")

# get a currencies' details
currencies_details <- kucoin::get_currency_details(c("BTC", "XMR"))

currencies_details

}
\seealso{
\code{kucoin::get_deposit_address()}
}
