% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market-data.R
\name{get_kucoin_symbols}
\alias{get_kucoin_symbols}
\title{Get all symbols' most recent metadata}
\usage{
get_kucoin_symbols()
}
\value{
A \code{tibble} containing some metadata
}
\description{
Get all symbols' most recent metadata
}
\examples{
# import library
library(kucoin)

# get all symbols' most recent metadata
metadata <- get_kucoin_symbols()

# quick check
metadata

}
