% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_data.R
\name{get_market_data}
\alias{get_market_data}
\title{Get historical market data for symbols}
\usage{
get_market_data(
  symbols = NULL,
  from = NULL,
  to = NULL,
  frequency = NULL,
  delay = 0,
  retries = 3
)
}
\arguments{
\item{symbols}{A `character` vector of one or more pair symbol (required - default `NULL`).}

\item{from}{A `character` with valid `date`/`datetime` format, or `date`/`datetime` object as a start of datetime range (required - default `NULL`).}

\item{to}{A `character` with valid `date`/`datetime` format, or `date`/`datetime` object as an end of datetime range (required - default `NULL`).}

\item{frequency}{A `character` vector of one which specify the frequency option, see details for further information (required - default `NULL`).}

\item{delay}{A `numeric` value to delay data request in milliseconds (optional - default `0`).}

\item{retries}{A `numeric` value to specify the number of retries in case of failure (optional - default `3`).}
}
\value{
A `data.table` with price data
}
\description{
Get historical market data for symbols
}
\details{
There are several supported frequencies:

 * `"1 minute"`
 * `"3 minutes"`
 * `"5 minutes"`
 * `"15 minutes"`
 * `"30 minutes"`
 * `"1 hour"`
 * `"2 hours"`
 * `"4 hours"`
 * `"6 hours"`
 * `"8 hours"`
 * `"12 hours"`
 * `"1 day"`
 * `"1 week"`

# ---------------
For more information see documentation: [KuCoin - get-klines](https://docs.kucoin.com/#get-klines)
}
\examples{

# get one symbol
kucoin::get_market_data(
    symbols = "BTC/USDT",
    from = "2022-11-05 00:00:00",
    to = "2022-11-06 00:00:00",
    frequency = "1 hour"
)

# get multiple symbols
kucoin::get_market_data(
    symbols = c("BTC/USDT", "XMR/BTC", "KCS/USDT"),
    from = "2022-11-05 00:00:00",
    to = "2022-11-06 00:00:00",
    frequency = "1 hour"
)

}
