% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get_market_metadata.R
\name{get_market_metadata.deprecated}
\alias{get_market_metadata.deprecated}
\title{Get all market symbols' metadata --deprecated--}
\usage{
get_market_metadata.deprecated(retries = 3)
}
\arguments{
\item{retries}{A `numeric` value to specify the number of retries in case of failure (optional - default `3`).}
}
\value{
A `data.table` with metadata
}
\description{
Get all market symbols' metadata --deprecated--
}
\details{
TODO: this function needs to be updated to v2 of the API.

# ---------------
For more information see documentation: [KuCoin - get-symbols-list-deprecated](https://docs.kucoin.com/#get-symbols-list-deprecated)
}
\examples{
# import library
library("kucoin")

# get all symbols' most recent metadata
metadata <- kucoin::get_market_metadata()

# quick check
metadata

}
