% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_metadata.R
\name{get_market_metadata}
\alias{get_market_metadata}
\title{Get all symbols' most recent metadata}
\usage{
get_market_metadata(retries = 3)
}
\arguments{
\item{retries}{A \code{numeric} value to specify the number of retries in case of failure.}
}
\value{
A \code{data.table} containing some metadata
}
\description{
Get all symbols' most recent metadata
}
\examples{
# import library
library("kucoin")

# get all symbols' most recent metadata
metadata <- kucoin::get_market_metadata()

# quick check
metadata

}
