% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orders_all.R
\name{get_orders_all}
\alias{get_orders_all}
\title{Get all order(s)}
\usage{
get_orders_all(
  symbol = NULL,
  status = NULL,
  side = NULL,
  type = NULL,
  tradeType = "TRADE",
  from = NULL,
  to = NULL,
  retries = 3,
  delay = 0.1
)
}
\arguments{
\item{symbol}{A \code{character} vector of one or more which contain symbol(s) of format "BTC/USDT" (optional - default \code{NULL}).}

\item{status}{A \code{character} vector of one either "done" or "active" (optional - default \code{NULL}).}

\item{side}{A \code{character} vector of one either "buy" or "sell" (optional - default \code{NULL}).}

\item{type}{A \code{character} vector of one either "limit", "market", "limit_stop", "market_stop" (optional - default \code{NULL}).}

\item{tradeType}{A \code{character} vector of one either "TRADE" (spot trading), "MARGIN_TRADE" (cross margin trading), "MARGIN_ISOLATED_TRADE" (isolated margin trading) (required - default \code{TRADE}).}

\item{from}{A \code{character} with valid \code{date}/\code{datetime} format, or \code{date}/\code{datetime} object as a start of datetime range (optional - default \code{NULL}).}

\item{to}{A \code{character} with valid \code{date}/\code{datetime} format, or \code{date}/\code{datetime} object as an end of datetime range (optional - default \code{NULL}).}

\item{retries}{A \code{numeric} value to specify the number of retries in case of failure (optional - default \code{3}).}

\item{delay}{A \code{numeric} value to delay data request in milliseconds (optional - default \code{0.1}).}
}
\value{
A \code{data.table} containing order details
}
\description{
Get all order(s)
}
\details{
For more information see documentation: \href{https://docs.kucoin.com/#list-orders}{KuCoin - list-orders}
}
\examples{

\dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

kucoin::get_orders_all()

}

}
