% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orders_by_id.R
\name{get_orders_by_id}
\alias{get_orders_by_id}
\title{Get order(s) details by order id}
\usage{
get_orders_by_id(orderIds = NULL, delay = 0, retries = 3)
}
\arguments{
\item{orderIds}{A \code{character} vector of one or more which contain the order id(s) designated by KuCoin (required - default \code{NULL}).}

\item{delay}{A \code{numeric} value to delay data request in milliseconds (optional - default \code{0}).}

\item{retries}{A \code{numeric} value to specify the number of retries in case of failure (optional - default \code{3}).}
}
\value{
A \code{data.table} containing order details
}
\description{
Get order(s) details by order id
}
\details{
For more information see documentation: \href{https://docs.kucoin.com/#get-an-order}{KuCoin - get-order}
}
\examples{

\dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

# get order details
order_details <- kucoin::get_orders_by_id(order_ids = "insertorderid")

# quick check
order_details

}

}
