% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_subaccount}
\alias{get_subaccount}
\title{Retrieve SubAccount Configuration from Environment Variables}
\usage{
get_subaccount(
  sub_account_name = Sys.getenv("KC-ACCOUNT-SUBACCOUNT-NAME"),
  sub_account_password = Sys.getenv("KC-ACCOUNT-SUBACCOUNT-PASSWORD")
)
}
\arguments{
\item{sub_account_name}{(Optional) The sub-account name. Defaults to \code{Sys.getenv("KC-ACCOUNT-SUBACCOUNT-NAME")}.}

\item{sub_account_password}{(Optional) The sub-account password. Defaults to \code{Sys.getenv("KC-ACCOUNT-SUBACCOUNT-PASSWORD")}.}
}
\value{
A list containing sub-account configuration parameters.
}
\description{
This function retrieves sub-account–specific configuration parameters from environment variables.
It expects the following environment variables to be set:
- \code{KC-ACCOUNT-SUBACCOUNT-NAME}: The sub-account name.
- \code{KC-ACCOUNT-SUBACCOUNT-PASSWORD}: The sub-account password.
}
\details{
These parameters are used for sub-account related operations.
}
\examples{
\dontrun{
  sub_cfg <- get_subaccount()
  print(sub_cfg)
}

}
