% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kucoin_time_to_datetime}
\alias{kucoin_time_to_datetime}
\title{Convert raw Kucoin time to a \code{datetime} object}
\usage{
kucoin_time_to_datetime(time)
}
\arguments{
\item{time}{A \code{numeric} vector of time returned from KuCoin API (milliseconds) to be converted to a \code{datetime} object.}
}
\value{
A \code{datetime} object
}
\description{
Convert raw Kucoin time to a \code{datetime} object
}
\examples{
# import library
library("kucoin")

# get current server time
kucoin_time_to_datetime(1.669401e+12)

}
