% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl_market_data_get_klines.R
\name{split_time_range_by_candles}
\alias{split_time_range_by_candles}
\title{Split a Time Range into Segments by Maximum Candle Count}
\usage{
split_time_range_by_candles(
  from,
  to,
  candle_duration_s,
  max_candles = 1500,
  overlap = 1
)
}
\arguments{
\item{from}{A POSIXct object representing the start time.}

\item{to}{A POSIXct object representing the end time.}

\item{candle_duration_s}{A numeric value; the duration of one candle in seconds.}

\item{max_candles}{Maximum number of candles per segment (default is 1500).}

\item{overlap}{Number of seconds to extend the end of each segment (default is 1).}
}
\value{
A data.table with two columns, `from` and `to`, where each row defines a segment.
}
\description{
Given a start and end time (POSIXct) and the duration of a single candle (in seconds),
this function splits the overall time range into segments such that each segment covers at most
`max_candles` candles.
}
\details{
Note: The function extends the end of each segment by `overlap` seconds to ensure that the last candle
when fetching data segments make sure to deduplicate the data.
}
\examples{
# For a 1min candle, a segment covers 1500 minutes.
from <- lubridate::now() - 2 * 3600
to <- lubridate::now()
candle_duration_s <- 60
split_time_range_by_candles(from, to, candle_duration_s)

}
