% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time_convert_kucoin.R
\name{time_convert_from_kucoin}
\alias{time_convert_from_kucoin}
\title{Convert KuCoin's Server Time to POSIXct}
\usage{
time_convert_from_kucoin(time_value, unit = c("ms", "ns", "s"))
}
\arguments{
\item{time_value}{A numeric value representing the time.}

\item{unit}{A character string specifying the time unit: "ms" for milliseconds (default) or "s" for seconds.}
}
\value{
A POSIXct object representing the time in UTC.
}
\description{
Converts a UNIX timestamp (in the specified unit) from KuCoin's server to a POSIXct object (UTC).
}
