% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_time_convert_kucoin.R
\name{time_convert_to_kucoin}
\alias{time_convert_to_kucoin}
\title{Convert POSIXct to KuCoin's Server Time}
\usage{
time_convert_to_kucoin(datetime, unit = c("ms", "ns", "s"))
}
\arguments{
\item{datetime}{A POSIXct object representing the time.}

\item{unit}{A character string specifying the output time unit: "ms" for milliseconds (default) or "s" for seconds.}
}
\value{
A numeric value representing the time in the specified unit.
}
\description{
Converts a POSIXct object to a UNIX timestamp in the specified unit for KuCoin's server.
}
