% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verify_symbol}
\alias{verify_symbol}
\title{Verify Symbol Format}
\usage{
verify_symbol(ticker)
}
\arguments{
\item{ticker}{A character string representing the symbol.}
}
\value{
A logical value; TRUE if the symbol is valid, FALSE otherwise.
}
\description{
Checks whether the ticker symbol is of the format "BTC-USDT" (uppercase alphanumeric separated by a dash).
}
